installer: move bitcoind steps to own module
